// apps/admin/src/api/generated/model/feedbackConversationDetailDtoOutputMessagesItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */
import type { FeedbackConversationDetailDtoOutputMessagesItemActor } from "./feedbackConversationDetailDtoOutputMessagesItemActor";
import type { FeedbackConversationDetailDtoOutputMessagesItemAttention } from "./feedbackConversationDetailDtoOutputMessagesItemAttention";
import type { FeedbackConversationDetailDtoOutputMessagesItemDelivery } from "./feedbackConversationDetailDtoOutputMessagesItemDelivery";

export type FeedbackConversationDetailDtoOutputMessagesItem = {
  actor: FeedbackConversationDetailDtoOutputMessagesItemActor;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  at: string;
  /** @nullable */
  attention: FeedbackConversationDetailDtoOutputMessagesItemAttention;
  /** @nullable */
  delivery: FeedbackConversationDetailDtoOutputMessagesItemDelivery;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  /**
   * @nullable
   * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  ingressId: string | null;
  /**
   * @nullable
   * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  outboxId: string | null;
  /**
   * @minLength 1
   * @maxLength 200
   * @nullable
   */
  providerMessageId: string | null;
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  seq: number;
  /**
   * @minLength 1
   * @maxLength 4096
   */
  text: string;
};